Module ini::ini

Ini

Structs

Ini
Ini struct

ParseError
Parse error

ParseOption
Parsing configuration

Properties
Properties type (key-value pairs)

SectionIter
Iterator for traversing sections

SectionIterMut
Iterator for traversing sections

SectionOccupiedEntry
A view into a occupied entry in a Ini

SectionSetter
A setter which could be used to set key-value pair in a specified section

SectionVacantEntry
A view into a vacant entry in a Ini

WriteOption
Writing configuration

Enums

Error
Error while parsing an INI document

EscapePolicy
Policies for escaping logic

LineSeparator
Newline style

SectionEntry
A view into an Ini, which may either be vacant or occupied.

Type Definitions

PropertyKey
Internal storage of property's key

SectionKey
Internal storage of section's key