Enum ini::ini::EscapePolicy

pub enum EscapePolicy {
    Nothing,
    Basics,
    BasicsUnicode,
    Reserved,
    ReservedUnicode,
    Everything,
}

Variants
Nothing

escape absolutely nothing (dangerous)

Basics

only escape the most necessary things

BasicsUnicode

escape basics and non-ascii characters

Reserved

Escape reserved symbols.

ReservedUnicode

Escape reserved symbols and non-ascii characters

Everything

Escape everything that some INI implementations assume

Methods

impl EscapePolicy

pub fn should_escape(&self, c: char) -> bool

Given a character this returns true if it should be escaped as per this policy or false if not.