Crate ini

Ini parser for Rust

use ini::Ini;

let mut conf = Ini::new();
conf.with_section(Some("User"))
    .set("name", "Raspberry树莓")
    .set("value", "Pi");
conf.with_section(Some("Library"))
    .set("name", "Sun Yat-sen U")
    .set("location", "Guangzhou=world");
conf.write_to_file("conf.ini").unwrap();

let i = Ini::load_from_file("conf.ini").unwrap();
for (sec, prop) in i.iter() {
    println!("Section: {:?}", sec);
    for (k, v) in prop.iter() {
        println!("{}:{}", k, v);
    }
}

Structs

Ini

Ini struct

ParseError

Parse error

ParseOption

Parsing configuration

Properties

Properties type (key-value pairs)

PropertiesIntoIter

PropertyIter

PropertyIterMut

Iterator for traversing sections

SectionIntoIter

Iterator for traversing sections

SectionIter

Iterator for traversing sections

SectionIterMut

Iterator for traversing sections

SectionOccupiedEntry

A view into a occupied entry in a Ini

SectionSetter

A setter which could be used to set key-value pair in a specified section

SectionVacantEntry

A view into a vacant entry in a Ini

WriteOption

Writing configuration

Enums

Error

Error while parsing an INI document

EscapePolicy

Policies for escaping logic

LineSeparator

Newline style

SectionEntry

A view into an Ini, which may either be vacant or occupied.

Type Aliases

PropertyKey

Internal storage of property’s key

SectionKey

Internal storage of section’s key