Struct ini::ini::Ini

pub struct Ini { /* fields omitted */ }

Ini struct

Methods

impl Ini

pub fn new() -> Ini

Create an instance

pub fn with_section<'b, S>(
&'b mut self,
section: Option<S>
) -> SectionSetter<'b> where
S: Into<String>,

Set with a specified section, None is for the general section

pub fn general_section(&self) -> &Properties

Get the immmutable general section

pub fn general_section_mut(&mut self) -> &mut Properties

Get the mutable general section

pub fn section<'a, S>(&'a self, name: Option<S>) -> Option<&'a Properties> where
S: Into<String>,

Get a immutable section

pub fn section_mut<'a, S>(
&'a mut self,
name: Option<S>
) -> Option<&'a mut Properties> where
S: Into<String>,

Get a mutable section

pub fn entry<'a>(
&'a mut self,
name: Option<String>
) -> Entry<Option<String>, Properties>

Get the entry

pub fn clear<'a>(&mut self)

Clear all entries

pub fn sections<'a>(&'a self) -> Keys<'a, Option<String>, Properties>

Iterate with sections

pub fn set_to<S>(&mut self, section: Option<S>, key: String, value: String) where
S: Into<String>,

Set key-value to a section

pub fn get_from<'a, S>(
&'a self,
section: Option<S>,
key: &str
) -> Option<&'a str> where
S: Into<String>,

Get the value from a section with key

Example:

use ini::Ini;
let input = "[sec]\nabc = def\n";
let ini = Ini::load_from_str(input).unwrap();
assert_eq!(ini.get_from(Some("sec"), "abc"), Some("def"));

pub fn get_from_or<'a, S>(
&'a self,
section: Option<S>,
key: &str,
default: &'a str
) -> &'a str where
S: Into<String>,

Get the value from a section with key, return the default value if it does not exist

Example:

use ini::Ini;
let input = "[sec]\n";
let ini = Ini::load_from_str(input).unwrap();
assert_eq!(ini.get_from_or(Some("sec"), "key", "default"), "default");

pub fn get_from_mut<'a, S>(
&'a mut self,
section: Option<S>,
key: &str
) -> Option<&'a str> where
S: Into<String>,

Get the mutable from a section with key

pub fn delete<S>(&mut self, section: Option<S>) -> Option<Properties> where
S: Into<String>,

Delete a section, return the properties if it exists

pub fn delete_from<S>(
&mut self,
section: Option<S>,
key: &str
) -> Option<String> where
S: Into<String>,

impl Ini

pub fn write_to_file<P: AsRef<Path>>(&self, filename: P) -> Result<()>

Write to a file

pub fn write_to_file_policy<P: AsRef<Path>>(
&self,
filename: P,
policy: EscapePolicy
) -> Result<()>

Write to a file

pub fn write_to<W: Write>(&self, writer: &mut W) -> Result<()>

Write to a writer

pub fn write_to_policy<W: Write>(
&self,
writer: &mut W,
policy: EscapePolicy
) -> Result<()>

Write to a writer

impl Ini

pub fn load_from_str(buf: &str) -> Result<Ini, Error>

Load from a string

pub fn load_from_str_noescape(buf: &str) -> Result<Ini, Error>

Load from a string, but do not interpret '' as an escape character

pub fn read_from<R: Read>(reader: &mut R) -> Result<Ini, Error>

Load from a reader

pub fn read_from_noescape<R: Read>(reader: &mut R) -> Result<Ini, Error>

Load from a reader, but do not interpret '' as an escape character

pub fn load_from_file<P: AsRef<Path>>(filename: P) -> Result<Ini, Error>

Load from a file

pub fn load_from_file_noescape<P: AsRef<Path>>(
filename: P
) -> Result<Ini, Error>

Load from a file, but do not interpret '' as an escape character

impl<'a> Ini

Important traits for SectionIterator<'a>
impl<'a> Iterator for SectionIterator<'a> type Item = (&'a Option<String>, &'a Properties);

pub fn iter(&'a self) -> SectionIterator<'a>

Immutable iterate though sections

Important traits for SectionMutIterator<'a>
impl<'a> Iterator for SectionMutIterator<'a> type Item = (&'a Option<String>, &'a mut Properties);

pub fn mut_iter(&'a mut self) -> SectionMutIterator<'a>

Mutable iterate though sections Deprecated! Use iter_mut instead!

Important traits for SectionMutIterator<'a>
impl<'a> Iterator for SectionMutIterator<'a> type Item = (&'a Option<String>, &'a mut Properties);

pub fn iter_mut(&'a mut self) -> SectionMutIterator<'a>

Mutable iterate though sections